Unnecessary casts - angbd is double, so ensures the rest of the calculation uses double
git-svn-id: https://svn.apache.org/repos/asf/commons/proper/math/trunk@1200350 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/commons/math/optimization/direct/BOBYQAOptimizer.java
@@ -2149,7 +2149,7 @@ public class BOBYQAOptimizer
             redsav = ZERO;
             iu = (int) (angbd * 17. + 3.1);
             for (int i = 0; i < iu; i++) {
-                angt = angbd * (double) i / (double) iu;
+                angt = angbd * i / iu;
                 sth = (angt + angt) / (ONE + angt * angt);
                 temp = shs + angt * (angt * dhd - dhs - dhs);
                 rednew = sth * (angt * dredg - sredg - HALF * sth * temp);
@@ -2171,7 +2171,7 @@ public class BOBYQAOptimizer
             }
             if (isav < iu) {
                 temp = (rdnext - rdprev) / (redmax + redmax - rdprev - rdnext);
-                angt = angbd * ((double) isav + HALF * temp) / (double) iu;
+                angt = angbd * (isav + HALF * temp) / iu;
             }
             cth = (ONE - angt * angt) / (ONE + angt * angt);
             sth = (angt + angt) / (ONE + angt * angt);
